better but backwards incompatible fix
Using the `bdist.format_command`  property is not possible and not required anymore. The property was removed in setuptools 63.4.0 (commit c0a4ea65ab68c8c6cb6ad54cc646c58bede6f0f4 on the setuptools repo).

So with setuptools 63.4.0 it is enough to update the `format_commands` dict, however this also requires setuptools 63.4.0 to work. If setuptools <63.4.0 is required, then the previous commit offers a solution.