Bugfix/incompatibility with setuptools 63.4.0 (#8)
* Constrain setuptools version to compatible versions

setuptools removed the `format_command` property of the `bdist` command, which setuptools-cythonize relies on.

This results in the following error:
```
File "D:\win_temp\pip-build-env-avucl34u\overlay\Lib\site-packages\setuptools_cythonize.py", line 166, in set_default_wheel_format
    cls.format_command['wheel'] = ('bdist_wheel', "Python .whl file")
AttributeError: type object 'CythonizeBdist' has no attribute 'format_command'
```

It works with setuptools versions <=63.3.0, so a workaround is to constrain the setuptools version to that.

* better but backwards incompatible fix

Using the `bdist.format_command`  property is not possible and not required anymore. The property was removed in setuptools 63.4.0 (commit c0a4ea65ab68c8c6cb6ad54cc646c58bede6f0f4 on the setuptools repo).

So with setuptools 63.4.0 it is enough to update the `format_commands` dict, however this also requires setuptools 63.4.0 to work. If setuptools <63.4.0 is required, then the previous commit offers a solution.

* Update setup.py

* Update setuptools_cythonize.py

Co-authored-by: Stefan Droege <droege@riscure.com>
Co-authored-by: Anxuae <anxuae-prog@yahoo.fr>